fix: code moves.
codes moved the general information component
