+++ src/pages/country-details/index.tsx
@@ -1,61 +1,8 @@
 import React from 'react'
-import { useRouter } from 'next/router'
-import { Box, Typography } from '@mui/material'
-import { useTheme } from 'src/context/themeContext'
-import useGetCountryDetails from 'src/services/hooks/useGetCountryDetail'
-import { DetailRow } from 'src/components/country-details/detail-row'
-import Chip from 'src/components/shared/chip'
-import { useTheme as useMuiTheme } from '@mui/material/styles'
+import { Box } from '@mui/material'
+import GeneralInformation from 'src/components/country-details/general-information'
 
 const CountryDetails = () => {
-  const router = useRouter()
-  const { countryName } = router.query
-  const { mode } = useTheme()
-  const theme = useMuiTheme()
-
-  const { data: countryData } = useGetCountryDetails({
-    filterValue: countryName as string,
-    fields:
-      'altSpellings,name,capital,population,languages,currencies,continents,region,subregion,area,landlocked,borders,maps,flags,coatOfArms,unMember,independent,timezones,startOfWeek,tld,postalCode'
-  })
-
-  if (!countryData) {
-    return <div>Loading...</div>
-  }
-
-  const generalInformationData = [
-    {
-      label: 'Official name',
-      value: countryData.name.official
-    },
-    {
-      label: 'Common name',
-      value: countryData.name.common
-    },
-    {
-      label: 'Alternative spellings',
-      value: countryData.altSpellings.join(', ')
-    },
-    {
-      label: 'Capital',
-      value: countryData.capital.join(', ')
-    },
-    {
-      label: 'Population',
-      value: countryData.population.toLocaleString()
-    },
-    {
-      label: 'Language',
-      value: Object.values(countryData.languages).join(', ')
-    },
-    {
-      label: 'Currencies',
-      value: Object.keys(countryData.currencies)
-        .map(code => `${countryData.currencies[code].name} (${countryData.currencies[code].symbol})`)
-        .join(', ')
-    }
-  ]
-
   return (
     <Box
       sx={{
@@ -66,30 +13,7 @@ const CountryDetails = () => {
         gap: '1rem'
       }}
     >
-      <Box sx={{ display: 'flex', alignItems: 'center', gap: '1.5rem' }}>
-        <Typography sx={{ fontWeight: 800, fontSize: '36px' }}>{countryData.name.common}</Typography>
-        <img src={countryData.flags.svg} alt={`${countryData.name.common} Flag`} height={24} width={36} />
-      </Box>
-      <Typography sx={{ fontWeight: 700, fontSize: '24px' }}>General Information</Typography>
-      <Box
-        sx={{
-          borderRadius: '8px',
-          border: `1px solid ${theme.palette.text.secondary}`,
-          bgcolor: mode === 'light' ? '#FFFFFF' : '#2B3743'
-        }}
-      >
-        {generalInformationData.map((data, index) => (
-          <Box key={index}>
-            {index === generalInformationData.length - 1 ? (
-              <DetailRow key={index} label={data.label} isLastMember>
-                <Chip color='blue' text={data.value} />
-              </DetailRow>
-            ) : (
-              <DetailRow key={index} label={data.label} value={data.value} isLastMember={false} />
-            )}
-          </Box>
-        ))}
-      </Box>
+      <GeneralInformation />
     </Box>
   )
 }
